requirements.txt
openai-whisper
sounddevice
scipy
requests
TTS
pvporcupine
pyaudio
python-dotenv
pvporcupine
pyaudio
whisper
